requirements.txt
numpy>=1.24,<2
Pillow>=10.0
PyYAML>=6.0
tqdm>=4.66
torch>=2.2
torchvision>=0.17
